Fixed frontend querys to db

+++ backend/controllers/students.js
@@ -31,25 +31,49 @@ export const getStudents = (req, res) => {
 }
 
 export const addStudent = (req, res) => {
-  new Student(req.body).save((err, doc) => {
-    if (err) {
-      res.status(500).send("Error adding student")
-    } else {
-      res.status(201).send("Created new student")
-    }
-  })
+  if (!req.body) {
+    res.status(400).send("Bad request")
+  } else {
+    new Student(req.body).save((err, doc) => {
+      if (err) {
+        console.error(err)
+        res.status(500).send("Error adding student")
+      } else {
+        res.status(201).send("Created new student")
+      }
+    })
+  }
+}
+
+export const updateStudentName = (req, res) => {
+  if (!req.body.name || !(req.body.email && req.body.name.firstName && req.body.name.lastName)) {
+    res.status(400).send("Bad request")
+  } else {
+    Student.findOneAndUpdate({ email: req.body.email }, { name: { firstName: req.body.name.firstName, lastName: req.body.name.lastName } }, (err, doc) => {
+      if (err) {
+        res.status(500).send("Error updating student information")
+      } else {
+        res.status(200).send("Successfully updated student information")
+      }
+    })
+  }
 }
 
 export const deleteStudent = (req, res) => {
-  Student.deleteOne({ email: req.body.email }, (err, doc) => {
-    if (err) {
-      console.err("Error deleting from database")
-    } else if (!doc) {
-      res.status(404).send("Email not found")
-    } else {
-      res.status(200).send("Successfully deleted")
-    }
-  })
+  if (!req.body.email) {
+    res.status(400).send("Bad request")
+  } else {
+    Student.deleteOne({ email: req.body.email }, (err, doc) => {
+      if (err) {
+        console.err("Error deleting from database")
+        res.status(500).send("Error deleting from database")
+      } else if (!doc) {
+        res.status(404).send("Email not found")
+      } else {
+        res.status(200).send("Successfully deleted")
+      }
+    })
+  }
 }
 
 export const deleteAllStudents = (req, res) => {
